refactor(library): move library components into ui/components/library and share the music player view model
