+++ app/src/main/java/io/github/aedev/flow/ui/components/AddToPlaylistDialog.kt
@@ -24,9 +24,9 @@ import androidx.compose.ui.unit.dp
 import coil3.compose.AsyncImage
 import io.github.aedev.flow.R
 import io.github.aedev.flow.data.local.PlaylistRepository
+import io.github.aedev.flow.data.model.PlaylistInfo
 import io.github.aedev.flow.data.model.Video
 import io.github.aedev.flow.ui.components.shared.rememberFlowSheetState
-import io.github.aedev.flow.ui.screens.playlists.PlaylistInfo
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.launch
