+++ app/src/main/java/io/github/aedev/flow/ui/components/library/PlaylistLibraryFilters.kt
@@ -1,4 +1,4 @@
-package io.github.aedev.flow.ui.screens.playlists
+package io.github.aedev.flow.ui.components.library
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
@@ -22,6 +22,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import io.github.aedev.flow.R
+import io.github.aedev.flow.data.model.PlaylistInfo
 import io.github.aedev.flow.ui.components.shared.FlowFilterChip
 import io.github.aedev.flow.ui.components.shared.MediaKind
