+++ app/src/main/java/io/github/aedev/flow/ui/components/music/sheet/MusicQuickActionsSheet.kt
@@ -35,7 +35,6 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
-import androidx.hilt.navigation.compose.hiltViewModel
 import io.github.aedev.flow.R
 import io.github.aedev.flow.data.music.model.MusicTrack
 import io.github.aedev.flow.ui.components.*
@@ -44,6 +43,7 @@ import io.github.aedev.flow.ui.components.music.sheet.AddToPlaylistDialog
 import io.github.aedev.flow.ui.components.music.sheet.CreatePlaylistDialog
 import io.github.aedev.flow.ui.components.shared.rememberFlowSheetState
 import io.github.aedev.flow.ui.screens.music.MusicPlayerViewModel
+import io.github.aedev.flow.ui.screens.music.sharedMusicPlayerViewModel
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
@@ -57,7 +57,7 @@ fun MusicQuickActionsSheet(
     onAudioEffectsClick: () -> Unit = {},
     onSleepTimerClick: () -> Unit = {},
     showPlaylistDialogs: Boolean = true,
-    viewModel: MusicPlayerViewModel = hiltViewModel(),
+    viewModel: MusicPlayerViewModel = sharedMusicPlayerViewModel(),
 ) {
     val context = androidx.compose.ui.platform.LocalContext.current
     val uiState by viewModel.uiState.collectAsState()
