+++ app/src/main/java/io/github/aedev/flow/ui/screens/home/HomeUiStateNormalization.kt
@@ -18,7 +18,7 @@ internal fun HomeUiState.withUniqueLazyContent(): HomeUiState {
         copy(
             videos = uniqueVideos,
             shorts = uniqueShorts,
-            continueWatchingVideos = uniqueHistory
+            continueWatchingVideos = uniqueHistory,
         )
     }
 }
