+++ app/src/main/java/io/github/aedev/flow/ui/screens/music/SharedMusicPlayerViewModel.kt
@@ -1,0 +1,12 @@
+package io.github.aedev.flow.ui.screens.music
+
+import androidx.activity.ComponentActivity
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.platform.LocalContext
+import androidx.hilt.navigation.compose.hiltViewModel
+
+@Composable
+fun sharedMusicPlayerViewModel(): MusicPlayerViewModel {
+    val activity = LocalContext.current as? ComponentActivity
+    return if (activity != null) hiltViewModel(activity) else hiltViewModel()
+}
