+++ app/src/test/java/io/github/aedev/flow/ui/screens/home/HomeFeedLogicTest.kt
@@ -17,21 +17,31 @@ import org.junit.Test
 
 /** I-10: behavioral coverage for the home-feed consumer logic (R-2 seeds, I-1 impressions). */
 class HomeFeedLogicTest {
-
-    private fun entry(id: String, pos: Long, dur: Long, ts: Long, isShort: Boolean = false) =
-        VideoHistoryEntry(
-            videoId = id, position = pos, duration = dur, timestamp = ts,
-            title = "t-$id", thumbnailUrl = "", isShort = isShort
-        )
+    private fun entry(
+        id: String,
+        pos: Long,
+        dur: Long,
+        ts: Long,
+        isShort: Boolean = false,
+    ) = VideoHistoryEntry(
+        videoId = id,
+        position = pos,
+        duration = dur,
+        timestamp = ts,
+        title = "t-$id",
+        thumbnailUrl = "",
+        isShort = isShort,
+    )
 
     @Test
     fun `history graph seeds keep non-shorts newest first with progress metadata`() {
-        val history = listOf(
-            entry("a", 90, 100, ts = 1),                  // 90% ✓
-            entry("b", 50, 100, ts = 2),                  // 50% ✗ below threshold
-            entry("c", 80, 100, ts = 3, isShort = true),  // short ✗
-            entry("d", 100, 100, ts = 4)                  // 100% ✓ newest
-        )
+        val history =
+            listOf(
+                entry("a", 90, 100, ts = 1), // 90% ✓
+                entry("b", 50, 100, ts = 2), // 50% ✗ below threshold
+                entry("c", 80, 100, ts = 3, isShort = true), // short ✗
+                entry("d", 100, 100, ts = 4), // 100% ✓ newest
+            )
         val seeds = graphSeedInputsFromHistory(history)
         assertThat(seeds.map { it.id }).containsExactly("d", "b", "a").inOrder()
         assertThat(seeds.first { it.id == "a" }.engagementWeight).isWithin(1e-6).of(0.9)
@@ -51,14 +61,25 @@ class HomeFeedLogicTest {
         assertThat(feedImpressionIds(visible, known)).containsExactly("v1", "v2").inOrder()
     }
 
-    private fun v(title: String, duration: Int) = Video(
-        id = title, title = title, channelName = "C", channelId = "c",
-        thumbnailUrl = "", duration = duration, viewCount = 1000, uploadDate = "1 day ago"
+    private fun v(
+        title: String,
+        duration: Int,
+    ) = Video(
+        id = title,
+        title = title,
+        channelName = "C",
+        channelId = "c",
+        thumbnailUrl = "",
+        duration = duration,
+        viewCount = 1000,
+        uploadDate = "1 day ago",
     )
 
-    private val defaultTaste = FeedTasteProfile(
-        comfortDurationSec = DURATION_COMFORT_DEFAULT_SEC, affinityTopics = emptySet()
-    )
+    private val defaultTaste =
+        FeedTasteProfile(
+            comfortDurationSec = DURATION_COMFORT_DEFAULT_SEC,
+            affinityTopics = emptySet(),
+        )
 
     @Test
     fun `format markers demote exploration content only for users with no matching interest`() {
@@ -75,18 +96,21 @@ class HomeFeedLogicTest {
         val essay = v("Three hour deep dive documentary", 11_000)
         assertThat(feedFitPenalty(essay, defaultTaste)).isGreaterThan(0.0)
         // Deep-diver / scholar tolerate any length: comfort cap is 0 ⇒ no duration penalty.
-        val deepDiver = feedTasteProfile(
-            UserBrain(totalInteractions = 200), FlowPersona.DEEP_DIVER
-        )
+        val deepDiver =
+            feedTasteProfile(
+                UserBrain(totalInteractions = 200),
+                FlowPersona.DEEP_DIVER,
+            )
         assertThat(feedFitPenalty(essay, deepDiver)).isEqualTo(0.0)
     }
 
     @Test
     fun `demoteByFit pushes poor-fit items below well-fit ones without dropping any`() {
-        val ranked = listOf(
-            v("3 Hour Baking Compilation Marathon", 11_000),   // marker + far over comfort
-            v("Kotlin coroutines explained", 900)              // clean
-        )
+        val ranked =
+            listOf(
+                v("3 Hour Baking Compilation Marathon", 11_000), // marker + far over comfort
+                v("Kotlin coroutines explained", 900), // clean
+            )
         val out = demoteByFit(ranked, defaultTaste)
         assertThat(out.map { it.id })
             .containsExactly("Kotlin coroutines explained", "3 Hour Baking Compilation Marathon")
@@ -98,12 +122,22 @@ class HomeFeedLogicTest {
     fun `well-fit content keeps its engine order`() {
         val ranked = listOf(v("Kotlin coroutines explained", 900), v("Python walkthrough", 1_200))
         assertThat(demoteByFit(ranked, defaultTaste).map { it.id })
-            .containsExactly("Kotlin coroutines explained", "Python walkthrough").inOrder()
+            .containsExactly("Kotlin coroutines explained", "Python walkthrough")
+            .inOrder()
     }
 
-    private fun vc(id: String, channelId: String) = Video(
-        id = id, title = id, channelName = channelId, channelId = channelId,
-        thumbnailUrl = "", duration = 600, viewCount = 1, uploadDate = "1 day ago"
+    private fun vc(
+        id: String,
+        channelId: String,
+    ) = Video(
+        id = id,
+        title = id,
+        channelName = channelId,
+        channelId = channelId,
+        thumbnailUrl = "",
+        duration = 600,
+        viewCount = 1,
+        uploadDate = "1 day ago",
     )
 
     private fun gc(
@@ -112,7 +146,7 @@ class HomeFeedLogicTest {
         seedScore: Double,
         graphRank: Int = 0,
         seedResultCount: Int = 12,
-        hitCount: Int = 1
+        hitCount: Int = 1,
     ) = GraphCandidate(
         video = video,
         seedId = seedId,
@@ -120,7 +154,7 @@ class HomeFeedLogicTest {
         graphRank = graphRank,
         seedCluster = "cluster-$seedId",
         seedResultCount = seedResultCount,
-        hitCount = hitCount
+        hitCount = hitCount,
     )
 
     private fun adjacentSameChannel(videos: List<Video>) =
@@ -128,9 +162,14 @@ class HomeFeedLogicTest {
 
     @Test
     fun `spaceByChannel separates clustered same-channel items without dropping any`() {
-        val clustered = listOf(
-            vc("a1", "A"), vc("a2", "A"), vc("a3", "A"), vc("b1", "B"), vc("c1", "C")
-        )
+        val clustered =
+            listOf(
+                vc("a1", "A"),
+                vc("a2", "A"),
+                vc("a3", "A"),
+                vc("b1", "B"),
+                vc("c1", "C"),
+            )
         val spaced = spaceByChannel(clustered)
         assertThat(spaced.map { it.id }).containsExactlyElementsIn(clustered.map { it.id })
         assertThat(adjacentSameChannel(spaced)).isEqualTo(0)
@@ -154,69 +193,90 @@ class HomeFeedLogicTest {
     fun `home feed quotas adapt to no subscription maturing and mature profiles`() {
         assertThat(homeFeedQuotas(40, subCount = 0, totalInteractions = 200))
             .containsExactly(
-                FeedSource.SUBS, 0,
-                FeedSource.RELATED, 14,
-                FeedSource.DISCOVERY, 18,
-                FeedSource.VIRAL, 8
+                FeedSource.SUBS,
+                0,
+                FeedSource.RELATED,
+                14,
+                FeedSource.DISCOVERY,
+                18,
+                FeedSource.VIRAL,
+                8,
             )
 
         assertThat(homeFeedQuotas(40, subCount = 12, totalInteractions = 20))
             .containsExactly(
-                FeedSource.SUBS, 14,
-                FeedSource.RELATED, 12,
-                FeedSource.DISCOVERY, 10,
-                FeedSource.VIRAL, 4
+                FeedSource.SUBS,
+                14,
+                FeedSource.RELATED,
+                12,
+                FeedSource.DISCOVERY,
+                10,
+                FeedSource.VIRAL,
+                4,
             )
 
         assertThat(homeFeedQuotas(40, subCount = 12, totalInteractions = 200))
             .containsExactly(
-                FeedSource.SUBS, 16,
-                FeedSource.RELATED, 10,
-                FeedSource.DISCOVERY, 10,
-                FeedSource.VIRAL, 4
+                FeedSource.SUBS,
+                16,
+                FeedSource.RELATED,
+                10,
+                FeedSource.DISCOVERY,
+                10,
+                FeedSource.VIRAL,
+                4,
             )
     }
 
     @Test
     fun `blendFeedSources reports source distribution`() {
-        val result = blendFeedSources(
-            lanes = mapOf(
-                FeedSource.SUBS to listOf(vc("s1", "S")),
-                FeedSource.RELATED to listOf(vc("r1", "R")),
-                FeedSource.DISCOVERY to listOf(vc("d1", "D")),
-                FeedSource.VIRAL to listOf(vc("v1", "V"))
-            ),
-            quotas = mapOf(
-                FeedSource.SUBS to 1,
-                FeedSource.RELATED to 1,
-                FeedSource.DISCOVERY to 1,
-                FeedSource.VIRAL to 1
-            ),
-            targetSize = 4
-        )
+        val result =
+            blendFeedSources(
+                lanes =
+                    mapOf(
+                        FeedSource.SUBS to listOf(vc("s1", "S")),
+                        FeedSource.RELATED to listOf(vc("r1", "R")),
+                        FeedSource.DISCOVERY to listOf(vc("d1", "D")),
+                        FeedSource.VIRAL to listOf(vc("v1", "V")),
+                    ),
+                quotas =
+                    mapOf(
+                        FeedSource.SUBS to 1,
+                        FeedSource.RELATED to 1,
+                        FeedSource.DISCOVERY to 1,
+                        FeedSource.VIRAL to 1,
+                    ),
+                targetSize = 4,
+            )
 
         assertThat(result.videos.map { it.id }).containsExactly("s1", "r1", "d1", "v1").inOrder()
         assertThat(result.sourceCounts)
             .containsExactly(
-                FeedSource.SUBS, 1,
-                FeedSource.RELATED, 1,
-                FeedSource.DISCOVERY, 1,
-                FeedSource.VIRAL, 1
+                FeedSource.SUBS,
+                1,
+                FeedSource.RELATED,
+                1,
+                FeedSource.DISCOVERY,
+                1,
+                FeedSource.VIRAL,
+                1,
             )
     }
 
     @Test
     fun `blendFeedSources relaxes scarce quota in related discovery subs viral order`() {
-        val result = blendFeedSources(
-            lanes = mapOf(
-                FeedSource.SUBS to listOf(vc("s1", "S"), vc("s2", "S")),
-                FeedSource.RELATED to listOf(vc("r1", "R1"), vc("r2", "R2")),
-                FeedSource.DISCOVERY to listOf(vc("d1", "D1"), vc("d2", "D2")),
-                FeedSource.VIRAL to listOf(vc("v1", "V1"), vc("v2", "V2"))
-            ),
-            quotas = FeedSource.entries.associateWith { 0 },
-            targetSize = 3
-        )
+        val result =
+            blendFeedSources(
+                lanes =
+                    mapOf(
+                        FeedSource.SUBS to listOf(vc("s1", "S"), vc("s2", "S")),
+                        FeedSource.RELATED to listOf(vc("r1", "R1"), vc("r2", "R2")),
+                        FeedSource.DISCOVERY to listOf(vc("d1", "D1"), vc("d2", "D2")),
+                        FeedSource.VIRAL to listOf(vc("v1", "V1"), vc("v2", "V2")),
+                    ),
+                quotas = FeedSource.entries.associateWith { 0 },
+                targetSize = 3,
+            )
 
         assertThat(result.items.map { it.source })
             .containsExactly(FeedSource.RELATED, FeedSource.RELATED, FeedSource.DISCOVERY)
@@ -245,13 +305,14 @@ class HomeFeedLogicTest {
         ranked[10] = oneOff
         ranked[11] = multiSeed
 
-        val boosted = applyGraphBoost(
-            ranked,
-            mapOf(
-                oneOff.id to gc(oneOff, seedId = "s1", seedScore = 0.4, graphRank = 0),
-                multiSeed.id to gc(multiSeed, seedId = "s2", seedScore = 1.2, graphRank = 0, hitCount = 2)
+        val boosted =
+            applyGraphBoost(
+                ranked,
+                mapOf(
+                    oneOff.id to gc(oneOff, seedId = "s1", seedScore = 0.4, graphRank = 0),
+                    multiSeed.id to gc(multiSeed, seedId = "s2", seedScore = 1.2, graphRank = 0, hitCount = 2),
+                ),
             )
-        )
 
         assertThat(boosted.indexOf(multiSeed)).isLessThan(boosted.indexOf(oneOff))
     }
@@ -264,10 +325,11 @@ class HomeFeedLogicTest {
         ranked[10] = clean
         ranked[11] = poorFit
 
-        val boosted = applyGraphBoost(
-            ranked,
-            mapOf(poorFit.id to gc(poorFit, seedId = "s1", seedScore = 1.2, graphRank = 0, hitCount = 2))
-        )
+        val boosted =
+            applyGraphBoost(
+                ranked,
+                mapOf(poorFit.id to gc(poorFit, seedId = "s1", seedScore = 1.2, graphRank = 0, hitCount = 2)),
+            )
         val demoted = demoteByFit(boosted, defaultTaste)
 
         assertThat(demoted.indexOf(clean)).isLessThan(demoted.indexOf(poorFit))
@@ -278,22 +340,24 @@ class HomeFeedLogicTest {
         val page = mutableListOf<Video>()
         val channelCounts = mutableMapOf<String, Int>()
         val usedIds = mutableSetOf("existing")
-        val candidates = listOf(
-            vc("existing", "A"),
-            vc("a1", "A"),
-            vc("a2", "A"),
-            vc("a3", "A"),
-            vc("b1", "B")
-        )
+        val candidates =
+            listOf(
+                vc("existing", "A"),
+                vc("a1", "A"),
+                vc("a2", "A"),
+                vc("a3", "A"),
+                vc("b1", "B"),
+            )
 
-        val added = addUniquePageVideos(
-            candidates = candidates,
-            targetList = page,
-            channelCounts = channelCounts,
-            usedVideoIds = usedIds,
-            targetSize = 3,
-            maxPerChannel = 2
-        )
+        val added =
+            addUniquePageVideos(
+                candidates = candidates,
+                targetList = page,
+                channelCounts = channelCounts,
+                usedVideoIds = usedIds,
+                targetSize = 3,
+                maxPerChannel = 2,
+            )
 
         assertThat(added).isEqualTo(3)
         assertThat(page.map { it.id }).containsExactly("a1", "a2", "b1").inOrder()
@@ -307,31 +371,35 @@ class HomeFeedLogicTest {
 
     @Test
     fun `related lane metrics summarize graph quality signals`() {
-        val seeds = listOf(
-            graphSeed("h1", GraphSeedSource.WATCH_HISTORY),
-            graphSeed("l1", GraphSeedSource.LIKED),
-            graphSeed("p1", GraphSeedSource.PLAYLIST)
-        )
+        val seeds =
+            listOf(
+                graphSeed("h1", GraphSeedSource.WATCH_HISTORY),
+                graphSeed("l1", GraphSeedSource.LIKED),
+                graphSeed("p1", GraphSeedSource.PLAYLIST),
+            )
         val r1 = vc("r1", "R")
         val r2 = vc("r2", "R")
-        val metrics = buildRelatedLaneMetrics(
-            seedInputs = seeds,
-            seedIds = listOf("h1", "l1"),
-            fetchedPerSeed = mapOf("h1" to 3, "l1" to 0),
-            mergedRelatedCandidates = listOf(gc(r1, "h1", 1.2), gc(r2, "h1", 1.1)),
-            filteredRelatedCandidates = listOf(gc(r1, "h1", 1.2)),
-            selectedSourceCounts = mapOf(
-                FeedSource.RELATED to 2,
-                FeedSource.DISCOVERY to 1,
-                FeedSource.SUBS to 1
-            ),
-            finalFeedCount = 4,
-            finalRelatedVideoIds = setOf("r1", "r2"),
-            brain = UserBrain(
-                watchHistoryMap = mapOf("r1" to 0.8f),
-                suppressedVideoIds = mapOf("r2" to 100L)
+        val metrics =
+            buildRelatedLaneMetrics(
+                seedInputs = seeds,
+                seedIds = listOf("h1", "l1"),
+                fetchedPerSeed = mapOf("h1" to 3, "l1" to 0),
+                mergedRelatedCandidates = listOf(gc(r1, "h1", 1.2), gc(r2, "h1", 1.1)),
+                filteredRelatedCandidates = listOf(gc(r1, "h1", 1.2)),
+                selectedSourceCounts =
+                    mapOf(
+                        FeedSource.RELATED to 2,
+                        FeedSource.DISCOVERY to 1,
+                        FeedSource.SUBS to 1,
+                    ),
+                finalFeedCount = 4,
+                finalRelatedVideoIds = setOf("r1", "r2"),
+                brain =
+                    UserBrain(
+                        watchHistoryMap = mapOf("r1" to 0.8f),
+                        suppressedVideoIds = mapOf("r2" to 100L),
+                    ),
             )
-        )
 
         assertThat(metrics.seedCandidatesAvailable).isEqualTo(3)
         assertThat(metrics.seedsSelected).isEqualTo(2)
@@ -349,7 +417,7 @@ class HomeFeedLogicTest {
     private fun graphSeed(
         id: String,
         source: GraphSeedSource,
-        timestamp: Long = 100L
+        timestamp: Long = 100L,
     ) = GraphSeedInput(
         id = id,
         title = "title-$id",
@@ -358,27 +426,29 @@ class HomeFeedLogicTest {
         engagementWeight = 1.0,
         timestamp = timestamp,
         durationSec = 600,
-        percentWatched = 100.0
+        percentWatched = 100.0,
     )
 
     @Test
     fun `saved interest seed inputs preserve source order before engine selection`() {
-        val sources = SavedSeedSources(
-            history = listOf(graphSeed("h_latest", GraphSeedSource.WATCH_HISTORY)),
-            liked = listOf(graphSeed("l1", GraphSeedSource.LIKED)),
-            playlists = listOf(graphSeed("p1", GraphSeedSource.PLAYLIST))
-        )
+        val sources =
+            SavedSeedSources(
+                history = listOf(graphSeed("h_latest", GraphSeedSource.WATCH_HISTORY)),
+                liked = listOf(graphSeed("l1", GraphSeedSource.LIKED)),
+                playlists = listOf(graphSeed("p1", GraphSeedSource.PLAYLIST)),
+            )
         val seeds = savedInterestSeedInputs(sources, cooldown = emptySet())
         assertThat(seeds.map { it.id }).containsExactly("h_latest", "l1", "p1").inOrder()
     }
 
     @Test
     fun `saved interest seed inputs pull from history liked and playlists within bounds`() {
-        val sources = SavedSeedSources(
-            history = (1..10).map { graphSeed("h$it", GraphSeedSource.WATCH_HISTORY) },
-            liked = (1..10).map { graphSeed("l$it", GraphSeedSource.LIKED) },
-            playlists = (1..10).map { graphSeed("p$it", GraphSeedSource.PLAYLIST) }
-        )
+        val sources =
+            SavedSeedSources(
+                history = (1..10).map { graphSeed("h$it", GraphSeedSource.WATCH_HISTORY) },
+                liked = (1..10).map { graphSeed("l$it", GraphSeedSource.LIKED) },
+                playlists = (1..10).map { graphSeed("p$it", GraphSeedSource.PLAYLIST) },
+            )
         val seeds = savedInterestSeedInputs(sources, cooldown = emptySet(), maxPerSource = 2)
         val ids = seeds.map { it.id }
         assertThat(seeds.size).isEqualTo(6)
@@ -389,14 +459,16 @@ class HomeFeedLogicTest {
 
     @Test
     fun `saved interest seed inputs exclude ids on cooldown`() {
-        val sources = SavedSeedSources(
-            history = listOf(
-                graphSeed("h_latest", GraphSeedSource.WATCH_HISTORY),
-                graphSeed("h2", GraphSeedSource.WATCH_HISTORY)
-            ),
-            liked = listOf(graphSeed("l1", GraphSeedSource.LIKED)),
-            playlists = listOf(graphSeed("p1", GraphSeedSource.PLAYLIST))
-        )
+        val sources =
+            SavedSeedSources(
+                history =
+                    listOf(
+                        graphSeed("h_latest", GraphSeedSource.WATCH_HISTORY),
+                        graphSeed("h2", GraphSeedSource.WATCH_HISTORY),
+                    ),
+                liked = listOf(graphSeed("l1", GraphSeedSource.LIKED)),
+                playlists = listOf(graphSeed("p1", GraphSeedSource.PLAYLIST)),
+            )
         val cooldown = setOf("h_latest", "l1", "p1")
         val seeds = savedInterestSeedInputs(sources, cooldown = cooldown)
         assertThat(seeds.map { it.id }).containsNoneIn(cooldown)
